added jitter lag to prevent ddos
